refactor(http): migrate structures and function definitions to http header file (#16) - Moved existing structures and new function definitions from `main.c` to a newly created `http.h` file. - This refactor improves modularity and makes it easier to manage HTTP-related data and functions in a separate header file.
